requirements_backup.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
tensorflow-cpu>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.25.0
joblib>=1.3.0
tqdm>=4.65.0
requests>=2.31.0